chore: Move predicate to utils
