+++ .golangci.yaml
@@ -131,6 +131,8 @@ linters-settings:
         alias: slicesutils
       - pkg: github.com/kyma-project/telemetry-manager/internal/utils/logger
         alias: loggerutils
+      - pkg: github.com/kyma-project/telemetry-manager/internal/utils/predicate
+        alias: predicateutils
   errcheck:
     check-type-assertions: true # Reports type assertions: `a := b.(SomeStruct)`.
     check-blank: true # Report assignment of errors to blank identifier: `num, _ := strconv.Atoi(numStr)`.
