+++ controllers/telemetry/logparser_controller.go
@@ -30,8 +30,8 @@ import (
 	"github.com/kyma-project/telemetry-manager/internal/conditions"
 	"github.com/kyma-project/telemetry-manager/internal/k8sutils"
 	"github.com/kyma-project/telemetry-manager/internal/overrides"
-	"github.com/kyma-project/telemetry-manager/internal/predicate"
 	"github.com/kyma-project/telemetry-manager/internal/reconciler/logparser"
+	predicateutils "github.com/kyma-project/telemetry-manager/internal/utils/predicate"
 	"github.com/kyma-project/telemetry-manager/internal/workloadstatus"
 )
 
@@ -77,6 +77,6 @@ func (r *LogParserController) SetupWithManager(mgr ctrl.Manager) error {
 		Watches(
 			&corev1.ConfigMap{},
 			handler.EnqueueRequestForOwner(mgr.GetClient().Scheme(), mgr.GetRESTMapper(), &telemetryv1alpha1.LogParser{}),
-			ctrlbuilder.WithPredicates(predicate.OwnedResourceChanged())).
+			ctrlbuilder.WithPredicates(predicateutils.OwnedResourceChanged())).
 		Complete(r)
 }
