+++ internal/utils/predicate/predicate.go
